+++ src/ClassicUO.Client/Configuration/Profile.cs
@@ -136,6 +136,7 @@ namespace ClassicUO.Configuration
         public bool EnableMusic { get; set => SetProperty(ref field, value); } = true;
         public int MusicVolume { get; set => SetProperty(ref field, value); } = 50;
         public bool EnableFootstepsSound { get; set => SetProperty(ref field, value); } = true;
+        public bool EnableRainSound { get; set => SetProperty(ref field, value); } = true;
         public bool EnableCombatMusic { get; set => SetProperty(ref field, value); } = true;
         public bool ReproduceSoundsInBackground { get; set => SetProperty(ref field, value); }
 
@@ -338,6 +339,7 @@ namespace ClassicUO.Configuration
         public bool PartyInviteGump { get; set => SetProperty(ref field, value); } = true;
         public bool CustomBarsToggled { get; set => SetProperty(ref field, value); }
         public bool CBBlackBGToggled { get; set => SetProperty(ref field, value); }
+        public bool UsePartyHealthBars { get; set => SetProperty(ref field, value); } = true;
 
         public bool ShowInfoBar { get; set => SetProperty(ref field, value); }
         public int InfoBarHighlightType { get; set => SetProperty(ref field, value); } // 0 = text colour changes, 1 = underline
@@ -367,6 +369,8 @@ namespace ClassicUO.Configuration
         public int AuraUnderFeetType { get; set => SetProperty(ref field, value); } // 0 = NO, 1 = in warmode, 2 = ctrl+shift, 3 = always
         public bool AuraOnMouse { get; set => SetProperty(ref field, value); } = true;
         public bool AnimatedWaterEffect { get; set => SetProperty(ref field, value); } = false;
+        public bool EnableWeatherEffects { get; set => SetProperty(ref field, value); } = false;
+        public bool EnableEnhancedWeather { get; set => SetProperty(ref field, value); } = false;
 
         public bool PartyAura { get; set => SetProperty(ref field, value); }
 
@@ -781,6 +785,8 @@ namespace ClassicUO.Configuration
         public double PaperdollScale { get; set => SetProperty(ref field, value); } = 1f;
 
         public double StatusGumpScale { get; set => SetProperty(ref field, Math.Clamp(value, 0.5d, 3.0d)); } = 1f;
+
+        public double ContextMenuScale { get; set => SetProperty(ref field, Math.Clamp(value, 0.5d, 3.0d)); } = 1f;
 
         /// <summary>
         /// Scale applied to every server created gump (and all of its controls).